requirements.txt
apache-airflow
polars
connectorx
sqlalchemy
python-dotenv
streamlit
requests
plotly
pandas
schedule
numpy
